backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
apturl==0.5.2
argcomplete==1.10.3
astroid==2.11.7
attrs==22.1.0
autopep8==1.7.0
Babel==2.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
compressed-rtf==1.0.6
croniter==0.3.29
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
dill==0.3.5.1
distlib==0.3.8
distro==1.4.0
distro-info==0.23+ubuntu1.1
docx2txt==0.8
ebcdic==1.1.1
entrypoints==0.3
exceptiongroup==1.2.0
extract-msg==0.28.7
Faker==19.13.0
fastapi==0.109.2
filelock==3.13.1
flake8==5.0.4
fsspec==2024.2.0
h11==0.14.0
httplib2==0.14.0
idna==2.8
IMAPClient==2.1.0
importlib-metadata==7.0.1
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
llvmlite==0.41.1
louis==3.12.0
lxml==4.9.1
macaroonbakery==1.3.1
MarkupSafe==2.1.5
mccabe==0.7.0
more-itertools==10.2.0
mpmath==1.3.0
netifaces==0.10.4
networkx==3.1
numba==0.58.1
numpy==1.24.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
openai-whisper==20231117
packaging==21.3
pandas==1.5.2
pdfminer.six==20191110
pexpect==4.6.0
Pillow==7.0.0
platformdirs==4.2.0
pluggy==1.0.0
protobuf==3.6.1
psutil==5.5.1
py==1.11.0
pycairo==1.16.2
pycodestyle==2.9.1
pycryptodome==3.15.0
pycups==1.9.73
pydantic==2.6.1
pydantic-core==2.16.2
pyflakes==2.5.0
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.14.5
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pytest==7.1.2
pytest-flake8==1.0.6
pytest-pylint==0.18.0
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-multipart==0.0.7
python-pptx==0.6.21
python-xlib==0.23
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
regex==2023.12.25
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
safeeyes==2.1.6
SecretStorage==2.3.1
simplejson==3.16.0
six==1.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
starlette==0.36.3
sympy==1.12
systemd-python==234
textract==1.6.5
tiktoken==0.6.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
torch==2.2.0
tqdm==4.66.1
triton==2.2.0
typing-extensions==4.9.0
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.27.0.post1
virtualenv==20.25.0
wadllib==1.3.3
wrapt==1.14.1
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.0.3
youtube-dl==2020.3.24
zipp==3.17.0
